Move componentoverwrites package into the model
