+++ pkg/components/model/componentoverwrites/componentoverwrites_suite_test.go
@@ -12,7 +12,7 @@ import (
 	. "github.com/onsi/gomega/gstruct"
 
 	lsv1alpha1 "github.com/gardener/landscaper/apis/core/v1alpha1"
-	"github.com/gardener/landscaper/pkg/landscaper/registry/componentoverwrites"
+	"github.com/gardener/landscaper/pkg/components/model/componentoverwrites"
 	testutils "github.com/gardener/landscaper/test/utils"
 )
