+++ pkg/components/model/componentversion_refs.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/gardener/landscaper/pkg/components/model/componentoverwrites"
 	"github.com/gardener/landscaper/pkg/components/model/types"
-	"github.com/gardener/landscaper/pkg/landscaper/registry/componentoverwrites"
 )
 
 // GetTransitiveComponentReferences returns a list of ComponentVersions that consists of the current one
